Prevent calls to dh_shlibdeps
Avoid calling dh_shlibdeps by hacking the PATH variable. A temporary
directory and a softlink to /bin/true are created. This directory is
pre-pended to the PATH when the call to alien is made. Therefore,
instead of alien calling the real dh_shlibdeps, it simply gets a null
output from true. Consequently, no library dependencies are generated,
avoiding the incorrect ones that may be generated if other packages are
installed.

Signed-off-by: Antonio Russo <antonio.e.russo@gmail.com>