Linux 4.18 compat: Kconfig
The kernel's top level Kconfig was updated in 4.18 such that it
no longer matches the expected pattern preventing the spl source
from being inserted.  Simplify the existing logic to always insert
the entry at the end of the file since an exact location is not
required, unlike in the Makefile.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue #7890